Move Event and Overlapped to talpid-windows crate
