+++ talpid-core/Cargo.toml
@@ -61,6 +61,7 @@ winreg = { version = "0.51", features = ["transactions"] }
 memoffset = "0.6"
 windows-service = "0.6.0"
 talpid-windows-net = { path = "../talpid-windows-net" }
+talpid-windows = { path = "../talpid-windows" }
 
 [target.'cfg(windows)'.dependencies.windows-sys]
 workspace = true
